GGB-1426 remove Float code
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51014 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ ggbjdk/src/main/java/org/geogebra/ggbjdk/java/awt/geom/Arc2D.java
@@ -322,7 +322,7 @@ public abstract class Arc2D extends RectangularShape implements GArc2D {
          */
         protected GRectangle2D makeBounds(double x, double y,
                                          double w, double h) {
-            return new Rectangle2D.Float((float) x, (float) y,
+            return new Rectangle2D.Double((float) x, (float) y,
                                          (float) w, (float) h);
         }
